replaced camera component with camera system.
